chore: dedupe drivers via _base, silence signal/filesystem hotspots
Tunnel drivers: extract the shared detect / spawn-loop / SIGTERM teardown
into _StreamProcessDriver; cloudflared / bore / ngrok now override only
the argv builder and the URL parser. Removes Sonar's new-code duplication
on these three files.

Signal hotspots: the two os.kill(pid, SIGTERM) call sites only target
PIDs we wrote into tunnels.json ourselves — NOSONAR'd with rationale.

Daemon tests: swap the /tmp placeholder workdir for /srv/bernstein so
Sonar's "publicly writable directory" hotspot doesn't fire on the test
fixtures.
